fix: rename shim and remove yarn dep
