+++ browser.js
