+++ package.json
@@ -47,8 +47,7 @@
     "react-navigation": "^3.11.1",
     "readable-stream": "^3.4.0",
     "unstated": "^2.1.1",
-    "vm-browserify": "1.1.0",
-    "yarn": "^1.17.3"
+    "vm-browserify": "1.1.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.5.0",
